Extract theme & update headers
