+++ src/App.tsx
@@ -1,20 +1,9 @@
 import React from "react";
 import Homepage from "./views/Homepage";
 import "./App.css";
-import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
+import { ThemeProvider } from "@material-ui/core/styles";
 
-const theme = createMuiTheme({
-  typography: {
-    fontFamily: ["Lato", "Garamond", "sans-serif"].join(","),
-  },
-  palette: {
-    primary: {
-      light: "#E2F4EC",
-      main: "#10844E",
-      dark: "#004B28",
-    },
-  },
-});
+import theme from "./theme";
 
 function App() {
   return (
